requirements.txt
streamlit
pandas
joblib
scikit-learn==1.3.1
google-generativeai